更新 COCO 數據集下載腳本，修正解壓縮步驟並調整資料型別以支持複數運算；新增 requirements.txt 檔案以列出依賴項

+++ OmniPose/dataset/coco.py
@@ -190,8 +190,8 @@ class COCODataset(JointsDataset):
             if max(obj['keypoints']) == 0:
                 continue
 
-            joints_3d = np.zeros((self.num_joints, 3), dtype=np.float)
-            joints_3d_vis = np.zeros((self.num_joints, 3), dtype=np.float)
+            joints_3d = np.zeros((self.num_joints, 3), dtype=np.cfloat)
+            joints_3d_vis = np.zeros((self.num_joints, 3), dtype=np.cfloat)
             for ipt in range(self.num_joints):
                 joints_3d[ipt, 0] = obj['keypoints'][ipt * 3 + 0]
                 joints_3d[ipt, 1] = obj['keypoints'][ipt * 3 + 1]
@@ -279,9 +279,9 @@ class COCODataset(JointsDataset):
             num_boxes = num_boxes + 1
 
             center, scale = self._box2cs(box)
-            joints_3d = np.zeros((self.num_joints, 3), dtype=np.float)
+            joints_3d = np.zeros((self.num_joints, 3), dtype=np.cfloat)
             joints_3d_vis = np.ones(
-                (self.num_joints, 3), dtype=np.float)
+                (self.num_joints, 3), dtype=np.cfloat)
             kpt_db.append({
                 'image': img_name,
                 'center': center,
